modify xpu_optimize_cachekv_initialization_pass to xpu_delete_cast_op_pass
